Render Dinosaur Head
Remove joints between edges and vertices to prevent physics from altering the Dino head skeleton and keep it clear to the user.

Add Awake() method to View.cs to start the Model script and register the listener at startup.

Store vertex and edge coordinates as a string in the Model to upload to the Quest.

Detect room boundaries to lay the Dino head naturally on the floor.

Clean up the code and remove unused parts.

Update edge and vertex prefabs.

Create DinosaurScene.

Make the Dino head interactable.

Use Logcat to check for runtime errors on Quest 3.